backend/pyproject.toml
[project]
name = "backend"
version = "0.3.4"
description = "FastAPI backend"
authors = [{ name = "EeHeng Chen", email = "author@example.com" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    "cryptography>=42.0.0",
    "anthropic>=0.40.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
    "pytest-mock",
    "ruff==0.14.0",
    "pytest-cov",
    "pytest-asyncio",
    "pre-commit",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# ONLY NEEDED IF WE WANT TO CREATE AN INSTALLABLE PACKAGE
# Find is relative to this pyproject.toml file.
# [tool.setuptools.packages.find]
# where = ["app"]
[tool.setuptools.packages.find]
include = ["app"]
exclude = ["alembic"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
# SQLAlchemy asyncio uses greenlets internally; without this, lines inside
# async route handlers that await DB calls appear uncovered.
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
target-version = "py312"
line-length = 79
cache-dir = "/tmp/ruff_cache"

[tool.ruff.lint]
exclude = ["tests"]
# SIM212: Checks for if expressions that check against a negated condition.
# SIM905: Consider using a list literal instead of `str.split`
ignore = ["C901", "F823", "SIM1", "SIM212", "SIM905"]
# RUF013: Checks for the use of implicit Optional
#  in type annotations when the default parameter value is None.
select = [
    "C",
    "E",
    "F",
    "I",
    "W",
    "RUF013",
    "UP006",
    "PERF102",
    "PLC1802",
    "PLC0208",
    "SIM",
]
extend-safe-fixes = ["UP006"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true
docstring-code-line-length = 79

[tool.ruff.lint.pycodestyle]
# E501 reports lines that exceed the length of 120.
max-line-length = 120
